Move MockTraverser to the tests.
Add ObjectFactory.getTraverser. FileNetObjectFactory returns a
DocumentTraverser, and FileNetProxies returns a MockTraverser.
