+++ src/com/google/enterprise/adaptor/filenet/ObjectFactory.java
@@ -23,6 +23,8 @@ import com.filenet.api.util.Id;
  * Factory for producing instances various FileNet Objects.
  */
 interface ObjectFactory {
+  FileNetAdaptor.Traverser getTraverser(ConfigOptions options);
+
   AutoConnection getConnection(String contentEngineUri,
       String username, String password)
       throws EngineRuntimeException;
